+++ examples/0.71/babel.config.js
@@ -1,3 +1,14 @@
 module.exports = {
   presets: ['module:metro-react-native-babel-preset'],
+  // plugins: [
+  //   [
+  //     'module-resolver',
+  //     {
+  //       alias: {
+  //         'react': require.resolve('react'),
+  //         'react-native': require.resolve('react-native'),
+  //       },
+  //     },
+  //   ],
+  // ],
 };
