+++ examples/0.71/metro.config.js
@@ -1,12 +1,11 @@
 const path = require('node:path');
 const escape = require('escape-string-regexp');
 const exclusionList = require('metro-config/src/defaults/exclusionList');
-const pak = require('../../package.json');
 
 const root = path.resolve(__dirname, '../..');
-
 const modules = Object.keys({
-  ...pak.peerDependencies,
+  ...require('../../package.json').peerDependencies,
+  ...require('../demo-kit/package.json').peerDependencies,
 });
 
 module.exports = {
@@ -16,12 +15,13 @@ module.exports = {
   // We need to make sure that only one version is loaded for peerDependencies
   // So we block them at the root, and alias them to the versions in example's node_modules
   resolver: {
-    blacklistRE: exclusionList(
-      modules.map(
+    blacklistRE: exclusionList([
+      ...modules.map(
         (m) =>
           new RegExp(`^${escape(path.join(root, 'node_modules', m))}\\/.*$`)
-      )
-    ),
+      ),
+      /demo-kit\/node_modules\/.*/,
+    ]),
 
     extraNodeModules: modules.reduce((acc, name) => {
       acc[name] = path.join(__dirname, 'node_modules', name);
