+++ examples/0.71/package.json
@@ -7,6 +7,7 @@
     "start": "react-native start"
   },
   "dependencies": {
+    "demo-kit": "workspace:^",
     "react": "18.2.0",
     "react-native": "0.71.19"
   },
