--- examples/0.71/src/components/index.ts
@@ -1,3 +1,0 @@
-export * from './Button';
-export * from './Section';
-export * from './EmptyText';
