+++ examples/demo-kit/README.md
@@ -1,0 +1,3 @@
+# demo-kit
+
+This is a demo kit for testing the react-native-zendesk-messaging package.
