+++ examples/demo-kit/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "demo-kit",
+  "private": true,
+  "main": "src/index.ts",
+  "devDependencies": {
+    "@types/react": "^18.2.6",
+    "react": "18.3.1",
+    "react-native": "0.77.0-rc.6"
+  },
+  "peerDependencies": {
+    "react": "*",
+    "react-native": "*"
+  }
+}
