+++ examples/demo-kit/src/components/EmptyText.tsx
@@ -2,13 +2,13 @@ import React from 'react';
 import { StyleSheet, Text } from 'react-native';
 import type { TextProps } from 'react-native';
 
-export const EmptyText = ({
-  children: text,
-}: TextProps): React.ReactElement => (
-  <Text style={[styles.text, text ? styles.blueText : styles.emptyText]}>
-    {text || '(Empty)'}
-  </Text>
-);
+export function EmptyText({ children: text }: TextProps): React.ReactElement {
+  return (
+    <Text style={[styles.text, text ? styles.blueText : styles.emptyText]}>
+      {text || '(Empty)'}
+    </Text>
+  );
+}
 
 const styles = StyleSheet.create({
   text: {
