+++ examples/demo-kit/src/components/Gap.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { StyleSheet, View } from 'react-native';
+
+export function Gap(): React.ReactElement {
+  return <View style={styles.gap} />;
+}
+
+const styles = StyleSheet.create({
+  gap: {
+    height: 16,
+  },
+});
